+++ code/utils/evaluation.py
@@ -7,42 +7,11 @@ import tensorflow as tf
 import numpy as np
 import settings
 from utils import measures
+from utils import classifiers
 
 
-class EnsembleClassifier:
-    """Wraps an ensemble model predicting a list of logits."""
-
-    def __init__(self, model):
-        self.model = model
-
-    def predict(self, x):
-        ensemble_logits = self.predict_logits(x)
-        ensemble_probs = []
-        for logits in ensemble_logits:
-            ensemble_probs.append(tf.nn.softmax(logits))
-        mean_probs = np.array(ensemble_probs).mean(axis=0)
-        return mean_probs
-
-    def predict_logits(self, x):
-        return self.model.predict(x)
-
-
-class IndividualClassifier:
-    """Wraps a model predicting logits."""
-
-    def __init__(self, model):
-        self.model = model
-
-    def predict(self, x):
-        logits = self.predict_logits(x)
-        probs = np.array(tf.nn.softmax(logits))
-        return probs
-
-    def predict_logits(self, x):
-        return self.model.predict(x)
-
-
-CLASSIFIER_WRAPPERS = {'ensemble': EnsembleClassifier, 'individual': IndividualClassifier}
+CLASSIFIER_WRAPPERS = {'ensemble': classifiers.EnsembleClassifier,
+                       'individual': classifiers.IndividualClassifier}
 
 
 def calc_classification_measures(model, images, labels, wrapper_type=None):
